Add branding to header and match to wire.
