+++ src/sass/base/_typography.scss
@@ -5,7 +5,7 @@ h4,
 h5,
 h6 {
   margin-top: 0;
-  font-family: $font-serif;
+  font-family: $font-sans;
 }
 
 body {
