+++ src/sass/components/header.scss
@@ -1,16 +1,27 @@
+@import '../styles';
+
 .site-header {
-  background: rebeccapurple;
+  background: top center $color-secondary;
+  border-top: 10px solid $color-primary-alt;
   margin-bottom: 1.45rem;
   h1 {
     margin: 0;
+
     a {
+      align-items: flex-start;
       color: white;
+      display: flex;
+      flex-direction: column;
       text-decoration: none;
+
+      img {
+        margin-bottom: .5em;
+      }
     }
   }
 }
 .site-header__inner {
   margin: 0 auto,;
-  max-width: 960px;
+  max-width: 95%;
   padding: 1.45rem 1.0875rem;
 }
