+++ src/sass/components/layout.scss
@@ -1,0 +1,5 @@
+// Helper styles for customizing the gatsby layout.
+
+body {
+  margin: 0;
+}
